Get utilities building under Linux + Clang
A number of tweaks were needed here:
1. Gcc/Clang x64 are LP64 compilers in that longs and pointers are 64-bit when compiling for that target platform. Windows is a P64 only. The problem here was that the types were marked as `long` in the part_record_type structure which meant they were 64-bit when compiling with Clang, so they have been changed to `int` which means they are regarded  as 32-bit. I may consider moving these types to fixed width types (i.e. `std::int32_t`) as long as they are being directly serialized in this way as it would be a more clear statement than the static_assert on size. As it stands, if we compiled on an ILP64 platform, we'd be in trouble as `int` would also be 64-bit.
2. Some additional headers are needed and modern clang is very good at holding us to account.
3. I had cheated a little with the boe_error, leveraging a sneaky detail of Visual Studio headers which provides an implementation of `what` which one can leverage. However that is not standard, so let's do it the right way.

+++ src/Windows Code Release 3/game_file_structures.hpp
@@ -344,7 +344,7 @@ typedef struct {
 static_assert(sizeof(outdoor_creature_type) == 30);
 
 typedef struct {
-	long age;
+	int age;
 	short gold, food;
 	unsigned char stuff_done[310][10], item_taken[200][8];
 	short light_level;
@@ -370,7 +370,7 @@ typedef struct {
 	short node_to_call[30];
 	char spec_items[50], help_received[120];
 	short m_killed[200];
-	long total_m_killed, total_dam_done, total_xp_gained, total_dam_taken;
+	int total_m_killed, total_dam_done, total_xp_gained, total_dam_taken;
 	char scen_name[256];
 } party_record_type;
 static_assert(sizeof(party_record_type) == 46400);
